Fixing translations stuff for command line tools and such.
We had switched mg_globals to have the default translations set to
None so that it would be set up by the mediagoblin app later.
However, this would mean that things like scripts would try to call
gettext and error out.

Thanks to Tumulte for catching this.

This commit sponsored by Aurimas Fišeras.  Thank you!