Always activate testing using pytest fixture.
This is possibly the cleaner approach to the problem solved
in the previous commit.

The previous commit makes sure, that testing is enabled at
import time. While this one enables it just before the
individual test is run. So this one takes a bit more CPU.

Maybe revert the previous commit?

References: 0536306048daa0970d2e43411ba2a9bf073e570e