Use unicode for logging comments
The comment problems detailed in issue 791 are related to logging
of comments creation. The log tries to format unicode comments into
an ascii string (that is the unicode comment content).

This also creates problems with mark seen functionality since that
also logs the comments which breaks and you end up with a lot of
international comments in your message queue.

This commit makes both log messages unicode as well as the
representation of the comment.