Added facilities for Pootle.
Previously we used Transifex as a web-tool for translations. Some time
ago Transifex went proprietary and we decided to move to Pootle.

Old script for updating translations was renamed and a new one was
added. Also a script for making the layout more Pootle-friendly was
added.

To accomplish the transition first run the old script:
./devtools/update_translations_transifex.sh

Then run the layout changing script:
./devtools/move_translations_for_pootle.sh

Commit and push. After that use the new
./devtools/update_translations_pootle.sh
to update the translations.

Closes bug: #913